+++ R/zzz.R
@@ -3,7 +3,7 @@
 # Detect repositories on package load.
 .onLoad <- function(libname, pkgname) {
   avail_on_init <- as.data.frame(available.packages(
-    repos = detect_repos(),
+    repos = detect_repos()
   ))[, c("Package", "Repository")]
 
   assign("avail_on_init", avail_on_init, envir = parent.env(environment()))